# swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CryptoPill",
		"contact": {},
		"version": "1.0",
		"description": "API Docs"
	},
	"servers": [
		{
			"url": "http://180.149.241.208:3048",
			"variables": {}
		},
		{
			"url": "http://localhost:3044",
			"variables": {}
		}
	],
	"paths": {
		"/laboratory/signup/getEmployeeDetails/{user_id}": {
			"get": {
				"tags": ["Laboratory"],
				"summary": "getEmployeeDetails",
				"description": "Fetch employee details",
				"operationId": "getEmployeeDetails",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"description": "",
						"required": true,
						"style": "simple",
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 3
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
		"/laboratory/signup/editEmployeeDetails": {
			"put": {
				"tags": ["Laboratory"],
				"summary": "editEmployeeDetails",
				"description": "To edit employee details in setup profile menu after a employee is added",
				"operationId": "editEmployeeDetails",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/editEmployeeDetailsRequest"
							},
							"example": {
								"id": 2,
								"first_name": "Ramo",
								"middle_name": "arest",
								"last_name": "pascal",
								"contact_number": "+916853672116",
								"gender": "male",
								"birth_date": "1990-07-11",
								"lab_or_pharma_employement_number": "1237",
								"document_number": "",
								"role_id": 4,
								"edit_profile": 0
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/laboratory/signup/addEmployeeDetails": {
			"post": {
				"tags": ["Laboratory"],
				"summary": "addEmployeeDetails",
				"description": "to add employee in lab set up profile for login purpose ",
				"operationId": "addEmployeeDetails",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/addEmployeeDetailsRequest"
							},
							"example": {
								"first_name": "Aslam",
								"middle_name": "abbas",
								"last_name": "Mansuir",
								"contact_number": "+918823672116",
								"gender": "male",
								"birth_date": "",
								"lab_or_pharma_employement_number": "1",
								"document_number": "1234ABCGHER",
								"document_type": "ADHAAR",
								"role_id": 4,
								"edit_profile": 1
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
		"/laboratory/updateLab": {
			"post": {
				"tags": ["Laboratory"],
				"summary": "updateSetupProfile",
				"operationId": "updateSetupProfile",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateSetupProfileRequest"
							},
							"example": {
								"laboratoryDetails": {
									"workplaces": [
										{
											"address": {
												"address_id": 5,
												"address": "Rama Tower, new Gopal Garden, Chembur - East , Mumbai, 400089",
												"location": {
													"latitude": 46.1672,
													"longitude": -46.353724
												},
												"locality": "sector 11",
												"city": "Mumbai",
												"pincode": "400341"
											},
											"workplace": {
												"workplace_id": 1,
												"workplace_name": "Jara Plutonics Laboratory",
												"is_franchise": 1,
												"franchise_name": "Jara Industries",
												"phone_number": "9898167634",
												"license_number": "AC129ETBMQ56U",
												"gst_number": "12845NKYTLDS0",
												"delivery_customer": 1,
												"delivery_distance": 10
											},
											"employees": [
												{
													"first_name": "lab_assistant4",
													"middle_name": null,
													"last_name": "lab_lastname4",
													"email": "testlab4@gmail.com",
													"gender": "male",
													"birth_date": "1995-11-07",
													"contact_number": "+919066660652",
													"lab_or_pharma_employement_number": "1102",
													"document_type": "ADHAAR",
													"document_number": "12FG1222040",
													"edit_profile": 1
												}
											]
										}
									],
									"user_id": 8
								},
								"user_role": 4
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/pharmacy/signup/removeEmployeeDetails/{user_id}": {
			"delete": {
				"tags": ["Pharmacy"],
				"summary": "Remove Employee",
				"description": "To remove pharmacy employee from setupprofile module",
				"operationId": "DeleteRemoveEmployee",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"description": "",
						"required": true,
						"style": "simple",
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 5
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/pharmacy/signup/getEmployeeDetails/{user_id}": {
			"get": {
				"tags": ["Pharmacy"],
				"summary": "getEmployeeDetails",
				"description": "Fetch employee details from Pharmacy setup profile module",
				"operationId": "GetgetEmployeeDetails",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"description": "",
						"required": true,
						"style": "simple",
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 5
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
		"/pharmacy/signup/addEmployeeDetails": {
			"post": {
				"tags": ["Pharmacy"],
				"summary": "addEmployeeDetails",
				"description": "To add employee for pharmacy is pharmacy setup profile module",
				"operationId": "PostaddEmployeeDetails",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/addEmployeeDetailsRequest"
							},
							"example": {
								"first_name": "dummy123",
								"middle_name": "",
								"last_name": "test123",
								"contact_number": "+918128153229",
								"gender": "male",
								"birth_date": "1985-11-07",
								"lab_or_pharma_employement_number": "1",
								"document_type": "ADHAAR",
								"document_number": "123EQRGDVGRF",
								"role_id": 5,
								"edit_profile": 0
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/pharmacy/signup/editEmployeeDetails": {
			"put": {
				"tags": ["Pharmacy"],
				"summary": "editEmployeeDetails",
				"description": "Edit pharmacy employee details in set up profile module",
				"operationId": "PuteditEmployeeDetails",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/editEmployeeDetailsRequest1"
							},
							"example": {
								"id": 4,
								"first_name": "Paras",
								"middle_name": "Mukesh",
								"last_name": "Shah",
								"contact_number": "+919418968163",
								"gender": "male",
								"birth_date": "1985-09-07",
								"lab_or_pharma_employement_number": "2",
								"document_type": "ADHAAR",
								"document_number": "125EQRADVGRFER",
								"role_id": 3,
								"edit_profile": 1
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/pharmacy/signup/addSetupProfileDetails": {
			"post": {
				"tags": ["Pharmacy"],
				"summary": "addSetupProfile",
				"operationId": "PostaddSetupProfile",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/addSetupProfileRequest1"
							},
							"example": {
								"pharmacyDetails": {
									"workplaces": [
										{
											"address": {
												"address": "Lokesh Tower, new Ambedkar Garden, Chembur - East , Mumbai, 400089",
												"location": {
													"latitude": 46.1082,
													"longitude": -46.363724
												},
												"locality": "sector 12",
												"city": "Mumbai",
												"pincode": "400341"
											},
											"workplace": {
												"workplace_name": "RSA Jandhan Pharmacy",
												"is_franchise": 1,
												"franchise_name": "RSA M Pharma Corps",
												"phone_number": "9898567634",
												"license_number": "AC129RTBMQ56U",
												"gst_number": "12845NKYLDS0",
												"delivery_customer": 1,
												"delivery_distance": 5,
												"minimum_order_amount": 100,
												"minimum_delivery_charges": 10,
												"additional_charges": 1,
												"minimum_additional_charges": 10
											}
										}
									]
								},
								"user_role": 5
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/pharmacy/updateSetupProfile": {
			"post": {
				"tags": ["Pharmacy"],
				"summary": "updateSetupProfile",
				"description": "UpdateSetupProfileDetails",
				"operationId": "PostupdateSetupProfile",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateSetupProfileRequest1"
							},
							"example": {
								"pharmacyDetails": {
									"workplaces": [
										{
											"address": {
												"address_id": 4,
												"address": "Natraj Complex, Plot No. CM 50 National Highway 8, Gunjan Rd",
												"location": {
													"latitude": 20.370843,
													"longitude": 72.922492
												},
												"locality": "GIDC",
												"city": "Vapi",
												"pincode": "396195"
											},
											"workplace": {
												"workplace_id": 2,
												"workplace_name": "21st Hospital local Pharmacy",
												"is_franchise": 1,
												"franchise_name": "21st Hospital Group",
												"phone_number": "+912602359317",
												"license_number": "AC129RTBMQ56U",
												"gst_number": "12845NKYLDS0",
												"delivery_customer": 1,
												"delivery_distance": 5,
												"minimum_order_amount": 500,
												"minimum_delivery_charges": 10,
												"additional_charges": 1,
												"minimum_additional_charges": 25
											},
											"employees": [
												{
													"first_name": "krishank",
													"middle_name": "",
													"last_name": "tripathi",
													"email": "krishank.tripathi@neosoftmail.com",
													"gender": "male",
													"birth_date": "1985-11-07",
													"contact_number": "+918128153228",
													"lab_or_pharma_employement_number": "1",
													"document_type": "ADHAAR",
													"document_number": "12FG12345678",
													"edit_profile": 1
												}
											]
										}
									]
								},
								"user_role": 5
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/user/resendOTP": {
			"post": {
			  "tags": [
				"User"
			  ],
			  "summary": "ResendOTP",
			  "operationId": "ResendOTP",
			  "parameters": [],
			  "requestBody": {
				"description": "",
				"content": {
				  "application/json": {
					"schema": {
					  "$ref": "#/components/schemas/ResendOTPRequest"
					},
					"example": {
					  "contact_number": "+918128153228"
					}
				  }
				},
				"required": true
			  },
			  "responses": {
				"200": {
				  "description": "",
				  "headers": {}
				}
			  },
			  "deprecated": false,
			  "security": []
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"editEmployeeDetailsRequest": {
				"title": "editEmployeeDetailsRequest",
				"required": [
					"id",
					"first_name",
					"middle_name",
					"last_name",
					"contact_number",
					"gender",
					"birth_date",
					"lab_or_pharma_employement_number",
					"document_number",
					"role_id",
					"edit_profile"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"first_name": {
						"type": "string"
					},
					"middle_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth_date": {
						"type": "string"
					},
					"lab_or_pharma_employement_number": {
						"type": "string"
					},
					"document_number": {
						"type": "string"
					},
					"role_id": {
						"type": "integer",
						"format": "int32"
					},
					"edit_profile": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"id": 2,
					"first_name": "Ramo",
					"middle_name": "arest",
					"last_name": "pascal",
					"contact_number": "+916853672116",
					"gender": "male",
					"birth_date": "1990-07-11",
					"lab_or_pharma_employement_number": "1237",
					"document_number": "",
					"role_id": 4,
					"edit_profile": 0
				}
			},
			"addEmployeeDetailsRequest": {
				"title": "addEmployeeDetailsRequest",
				"required": [
					"first_name",
					"middle_name",
					"last_name",
					"contact_number",
					"gender",
					"birth_date",
					"lab_or_pharma_employement_number",
					"document_number",
					"document_type",
					"role_id",
					"edit_profile"
				],
				"type": "object",
				"properties": {
					"first_name": {
						"type": "string"
					},
					"middle_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth_date": {
						"type": "string"
					},
					"lab_or_pharma_employement_number": {
						"type": "string"
					},
					"document_number": {
						"type": "string"
					},
					"document_type": {
						"type": "string"
					},
					"role_id": {
						"type": "integer",
						"format": "int32"
					},
					"edit_profile": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"first_name": "Aslam",
					"middle_name": "abbas",
					"last_name": "Mansuir",
					"contact_number": "+918823672116",
					"gender": "male",
					"birth_date": "",
					"lab_or_pharma_employement_number": "1",
					"document_number": "1234ABCGHER",
					"document_type": "ADHAAR",
					"role_id": 4,
					"edit_profile": 1
				}
			},
...
			"updateSetupProfileRequest": {
				"title": "updateSetupProfileRequest",
				"required": ["laboratoryDetails", "user_role"],
				"type": "object",
				"properties": {
					"laboratoryDetails": {
						"$ref": "#/components/schemas/LaboratoryDetails"
					},
					"user_role": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"laboratoryDetails": {
						"workplaces": [
							{
								"address": {
									"address_id": 5,
									"address": "Rama Tower, new Gopal Garden, Chembur - East , Mumbai, 400089",
									"location": {
										"latitude": 46.1672,
										"longitude": -46.353724
									},
									"locality": "sector 11",
									"city": "Mumbai",
									"pincode": "400341"
								},
								"workplace": {
									"workplace_id": 1,
									"workplace_name": "Jara Plutonics Laboratory",
									"is_franchise": 1,
									"franchise_name": "Jara Industries",
									"phone_number": "9898167634",
									"license_number": "AC129ETBMQ56U",
									"gst_number": "12845NKYTLDS0",
									"delivery_customer": 1,
									"delivery_distance": 10
								},
								"employees": [
									{
										"first_name": "lab_assistant4",
										"middle_name": null,
										"last_name": "lab_lastname4",
										"email": "testlab4@gmail.com",
										"gender": "male",
										"birth_date": "1995-11-07",
										"contact_number": "+919066660652",
										"lab_or_pharma_employement_number": "1102",
										"document_type": "ADHAAR",
										"document_number": "12FG1222040",
										"edit_profile": 1
									}
								]
							}
						],
						"user_id": 8
					},
					"user_role": 4
				}
			},
			"LaboratoryDetails": {
				"title": "LaboratoryDetails",
				"required": ["workplaces", "user_id"],
				"type": "object",
				"properties": {
					"workplaces": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Workplace1"
						},
						"description": ""
					},
					"user_id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"workplaces": [
						{
							"address": {
								"address_id": 5,
								"address": "Rama Tower, new Gopal Garden, Chembur - East , Mumbai, 400089",
								"location": {
									"latitude": 46.1672,
									"longitude": -46.353724
								},
								"locality": "sector 11",
								"city": "Mumbai",
								"pincode": "400341"
							},
							"workplace": {
								"workplace_id": 1,
								"workplace_name": "Jara Plutonics Laboratory",
								"is_franchise": 1,
								"franchise_name": "Jara Industries",
								"phone_number": "9898167634",
								"license_number": "AC129ETBMQ56U",
								"gst_number": "12845NKYTLDS0",
								"delivery_customer": 1,
								"delivery_distance": 10
							},
							"employees": [
								{
									"first_name": "lab_assistant4",
									"middle_name": null,
									"last_name": "lab_lastname4",
									"email": "testlab4@gmail.com",
									"gender": "male",
									"birth_date": "1995-11-07",
									"contact_number": "+919066660652",
									"lab_or_pharma_employement_number": "1102",
									"document_type": "ADHAAR",
									"document_number": "12FG1222040",
									"edit_profile": 1
								}
							]
						}
					],
					"user_id": 8
				}
			},
...
			"editEmployeeDetailsRequest1": {
				"title": "editEmployeeDetailsRequest1",
				"required": [
					"id",
					"first_name",
					"middle_name",
					"last_name",
					"contact_number",
					"gender",
					"birth_date",
					"lab_or_pharma_employement_number",
					"document_type",
					"document_number",
					"role_id",
					"edit_profile"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"first_name": {
						"type": "string"
					},
					"middle_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth_date": {
						"type": "string"
					},
					"lab_or_pharma_employement_number": {
						"type": "string"
					},
					"document_type": {
						"type": "string"
					},
					"document_number": {
						"type": "string"
					},
					"role_id": {
						"type": "integer",
						"format": "int32"
					},
					"edit_profile": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"id": 4,
					"first_name": "Paras",
					"middle_name": "Mukesh",
					"last_name": "Shah",
					"contact_number": "+919418968163",
					"gender": "male",
					"birth_date": "1985-09-07",
					"lab_or_pharma_employement_number": "2",
					"document_type": "ADHAAR",
					"document_number": "125EQRADVGRFER",
					"role_id": 3,
					"edit_profile": 1
				}
			},
			"addSetupProfileRequest1": {
				"title": "addSetupProfileRequest1",
				"required": ["pharmacyDetails", "user_role"],
				"type": "object",
				"properties": {
					"pharmacyDetails": {
						"$ref": "#/components/schemas/PharmacyDetails"
					},
					"user_role": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"pharmacyDetails": {
						"workplaces": [
							{
								"address": {
									"address": "Lokesh Tower, new Ambedkar Garden, Chembur - East , Mumbai, 400089",
									"location": {
										"latitude": 46.1082,
										"longitude": -46.363724
									},
									"locality": "sector 12",
									"city": "Mumbai",
									"pincode": "400341"
								},
								"workplace": {
									"workplace_name": "RSA Jandhan Pharmacy",
									"is_franchise": 1,
									"franchise_name": "RSA M Pharma Corps",
									"phone_number": "9898567634",
									"license_number": "AC129RTBMQ56U",
									"gst_number": "12845NKYLDS0",
									"delivery_customer": 1,
									"delivery_distance": 5,
									"minimum_order_amount": 100,
									"minimum_delivery_charges": 10,
									"additional_charges": 1,
									"minimum_additional_charges": 10
								}
							}
						]
					},
					"user_role": 5
				}
			},
			"PharmacyDetails": {
				"title": "PharmacyDetails",
				"required": ["workplaces"],
				"type": "object",
				"properties": {
					"workplaces": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Workplace6"
						},
						"description": ""
					}
				},
				"example": {
					"workplaces": [
						{
							"address": {
								"address": "Lokesh Tower, new Ambedkar Garden, Chembur - East , Mumbai, 400089",
								"location": {
									"latitude": 46.1082,
									"longitude": -46.363724
								},
								"locality": "sector 12",
								"city": "Mumbai",
								"pincode": "400341"
							},
							"workplace": {
								"workplace_name": "RSA Jandhan Pharmacy",
								"is_franchise": 1,
								"franchise_name": "RSA M Pharma Corps",
								"phone_number": "9898567634",
								"license_number": "AC129RTBMQ56U",
								"gst_number": "12845NKYLDS0",
								"delivery_customer": 1,
								"delivery_distance": 5,
								"minimum_order_amount": 100,
								"minimum_delivery_charges": 10,
								"additional_charges": 1,
								"minimum_additional_charges": 10
							}
						}
					]
				}
			},
...
			"updateSetupProfileRequest1": {
				"title": "updateSetupProfileRequest1",
				"required": ["pharmacyDetails", "user_role"],
				"type": "object",
				"properties": {
					"pharmacyDetails": {
						"$ref": "#/components/schemas/PharmacyDetails1"
					},
					"user_role": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"pharmacyDetails": {
						"workplaces": [
							{
								"address": {
									"address_id": 4,
									"address": "Natraj Complex, Plot No. CM 50 National Highway 8, Gunjan Rd",
									"location": {
										"latitude": 20.370843,
										"longitude": 72.922492
									},
									"locality": "GIDC",
									"city": "Vapi",
									"pincode": "396195"
								},
								"workplace": {
									"workplace_id": 2,
									"workplace_name": "21st Hospital local Pharmacy",
									"is_franchise": 1,
									"franchise_name": "21st Hospital Group",
									"phone_number": "+912602359317",
									"license_number": "AC129RTBMQ56U",
									"gst_number": "12845NKYLDS0",
									"delivery_customer": 1,
									"delivery_distance": 5,
									"minimum_order_amount": 500,
									"minimum_delivery_charges": 10,
									"additional_charges": 1,
									"minimum_additional_charges": 25
								},
								"employees": [
									{
										"first_name": "krishank",
										"middle_name": "",
										"last_name": "tripathi",
										"email": "krishank.tripathi@neosoftmail.com",
										"gender": "male",
										"birth_date": "1985-11-07",
										"contact_number": "+918128153228",
										"lab_or_pharma_employement_number": "1",
										"document_type": "ADHAAR",
										"document_number": "12FG12345678",
										"edit_profile": 1
									}
								]
							}
						]
					},
					"user_role": 5
				}
			},
...
			"ResendOTPRequest": {
				"title": "ResendOTPRequest",
				"required": [
				  "contact_number"
				],
				"type": "object",
				"properties": {
				  "contact_number": {
					"type": "string"
				  }
				},
				"example": {
				  "contact_number": "+918128153228"
				}
			}
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"security": [
		{
			"ApiKeyAuth": []
		}
	]
}